==> Project_5/01_Arabic_Data_Cleaning.ipynb <==
import pandas as pd
import numpy as np
import csv

import pickle

from langdetect import detect_langs, detect
from html.parser import HTMLParser
import re
import string

file_list = ['03_20','04_20', '05_20', '06_20', '07_20', '08_20', '09_20', '10_20', '11_20','12_20','01_21','02_21', '22_03_20', '2_04_20', '2_05_20', '2_06_20', '2_07_20', '2_08_20', '2_09_20', '2_10_20', '2_11_20', '2_12_20', '2_01_21','2_02_21']

url = '/Users/jess/workspace/twint/twint/arabic_tweets{}.csv'

def get_data(file_list, url, sep):
    df_list = []
    for file in file_list:
        file_url = url.format(file)
        df_list.append(pd.read_csv(file_url, sep=sep))
    return pd.concat(df_list)

init_df = get_data(file_list, url, sep='\t')

init_df.head()

init_df.shape

init_df.sort_values(by='date', inplace=True, ascending=True)

init_df.head()

init_df.describe()

init_df.drop(columns=['near', 'geo', 'source', 'user_rt_id', 'user_rt', 'retweet_id', 'retweet_date', 'translate', 'trans_src', 'trans_dest'], inplace=True)

init_df.columns

print(init_df['date'].max(), init_df['date'].min())

def get_value_counts(df):
    for colname, colval in df.iteritems():
        print(colname, colval.value_counts())

get_value_counts(init_df)

detect_langs(init_df['tweet'].iloc[0])

type(init_df['tweet'].iloc[0])

init_df['text'].iloc[0]

def clean_text(text):
    text = HTMLParser().unescape(text)
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
    text = re.sub('\w*\d\w*', '', text)
    return text
    

init_df['text'] = init_df.tweet.apply(clean_text)

init_df[['tweet', 'text']].head()

# init_df['lang'] = init_df['text'].map(lambda x: detect(x))

def detect_lang(text):
    try:
        lang = detect(text)
        return lang
    except:
        return "NaN"
        
    

init_df['lang'] = init_df['text'].apply(detect_lang)

init_df.lang.value_counts()

init_df['lang'].isnull().sum()

ar_df = init_df[init_df['lang']=='ar']

init_df.shape

ar_df.shape

def clean_text(text):
#     text = HTMLParser().unescape(text)
    text = re.sub(pattern='https?://\S+|www\.\S+', repl=' ', string=text)
    text = re.sub('[%s]' % re.escape(string.punctuation), ' ', text)
    text = re.sub('\w*\d\w*', ' ', text)
    return text
    

url_pattern = r'https?://\S+|www\.\S+'
text = re.sub(pattern=url_pattern, repl=' ', string=text)
return without_urls

ar_df['text'] = ar_df.tweet.apply(clean_text)

ar_df[['tweet', 'text']].head()

ar_df.to_pickle("ar_df.pkl")

ar_df[['tweet', 'text']].sample(25, random_state=55)

ar_df.lang.value_counts()
