==> Project_5/03_EDA_Preprocessing_AR.ipynb <==
import numpy as np
import pandas as pd
import pickle

from sklearn.feature_extraction.text import TfidfVectorizer, TfidfTransformer

from sklearn.decomposition import NMF
from sklearn.metrics.pairwise import cosine_similarity

from gensim import corpora, models, similarities, matutils

import linguistica as lxa
from collections import Counter

ar_df = pickle.load(open("ar_df.pkl", "rb"))

from camel_tools.tokenizers.word import simple_word_tokenize
from camel_tools.utils.normalize import normalize_unicode
from camel_tools.dialectid import DialectIdentifier, label_to_region

from camel_tools.disambig.mle import MLEDisambiguator
from camel_tools.tokenizers.morphological import MorphologicalTokenizer
from camel_tools.tagger.default import DefaultTagger

def clean_ar_text(text):
    mle_msa = MLEDisambiguator.pretrained('calima-msa-r13')
    tokenizer = MorphologicalTokenizer(disambiguator=mle_msa, 
                                       scheme='bwtok',
                                      split=True)
    
    norm_text = normalize_unicode(text)
    clean_text = norm_text.strip()
    simple_tokens = simple_word_tokenize(clean_text)
    tokens = tokenizer.tokenize(simple_tokens)
    tokens_clean = []
    for token in tokens:
        if "+" not in token:
            tokens_clean.append(token)
    return tokens_clean

ar_df[['text', 'tweet']].iloc[0]

len(ar_df['tweet'].iloc[0])

# text = ar_df['text'][0].apply(clean_ar_text)

type(ar_df['text'].iloc[0])

token = normalize_unicode(ar_df['text'].iloc[0])

token.strip()

tokens = simple_word_tokenize(token)

mle_msa = MLEDisambiguator.pretrained('calima-msa-r13')
tokenizer = MorphologicalTokenizer(disambiguator=mle_msa, 
                                       scheme='bwtok',
                                      split=True)

tokens = tokenizer.tokenize(tokens)

tokens

clean_ar_text(ar_df['text'].iloc[0])

len(text)

print(text)

print(ar_df['text'].iloc[0])

ar_df['tokens'] = ar_df['text'].apply(clean_ar_text)

# check tokens ^ this cell run recently

# ar_df[['tweet', 'text', 'tokens']].head()

ar_df.lang.value_counts()

# def id_dialect(text):
#     DID = DialectIdentifier.pretrained()
#     predictions = DID.predict(text)
# #     region_id = label_to_region(DID.predict(text))
#     top_dialect = [p.top for p in predictions]
#     return top_dialect
    

# text = ar_df['clean_text'][0:5].apply(id_dialect)

def p_o_s(text):
    mled = MLEDisambiguator.pretrained()
    tagger = DefaultTagger(mled, 'pos')
    pos_tagged = tagger.tag(text.split())
    return zip(text, pos_tagged)

ar_df['POS'] = ar_df['tokens'].apply(p_o_s)

def nouns(word, pos):
    is_noun = lambda pos: pos[:1] == 'noun'
    all_nouns = [word for (word, pos) in text if is_noun(text)]
    return all_nouns

ar_df['nouns'] = ar_df['POS'].apply(lambda (word, pos): nouns(word, pos))

ar_df[['tokens','nouns', 'POS']].head()

pickle.dump(ar_df, open('ar_df.pkl', 'wb'))

from sklearn.feature_extraction import text


stopword_url = "/Users/jess/workspace/Metis_Projects/Project_5/camel_tools/arabic-stop-words/list.txt"

def get_stopwords(file_location):
    lines = []
    file = open(file_location)
    lines = file.read().split('\n')
        
    file.close()
    return lines

stopwords = get_stopwords(stopword_url)

stopwords[0:5]

# my_stop_words = text.ENGLISH_STOP_WORDS.union(stopwords)
my_stop_words = set(text.ENGLISH_STOP_WORDS + stopwords + 'http' + 'https')

len(my_stop_words)

ar_df['filtered'] = en_df['tokens'].apply(lambda tweet: [word for word in tweet if word not in my_stop_words])

vectorizer = TfidfVectorizer(min_df=2, 
                             stop_words=my_stop_words)
tfidf1 = vectorizer.fit_transform(ar_df['filtered'])

vocab = vectorizer.vocabulary_

len(vocab)

idf = vectorizer.idf_

type(idf)

idf[0:5]

ar_tfidf_df = pd.DataFrame(tfidf1.toarray(), columns=vectorizer.get_feature_names())

ar_tfidf_df.head()

pickle.dump(tfidf1, open("tfidf1_sparse.pkl", 'wb'))
pickle.dump(vectorizer, open('ar_vectorizer.pkl', 'wb'))

def get_ngrams(df):
    data_dict = dict()

    tokens = []
    for row in en_df.index.tolist():
        tokens += en_df.loc[row, 'filtered'].split()

    lxa_object = lxa.from_corpus(tokens)

    unigrams = lxa_object.word_unigram_counter()
    bigrams = lxa_object.word_bigram_counter()
    trigrams = lxa_object.word_trigram_counter()

    df_unigrams = pd.DataFrame.from_dict(unigrams, orient='index') \
    .reset_index()


    df_bigrams = pd.DataFrame.from_dict(bigrams, orient='index') \
    .reset_index()


    df_trigrams = pd.DataFrame.from_dict(trigrams, orient='index') \
    .reset_index()

    df_unigrams.columns = ['N-Gram','Frequency']
    df_bigrams.columns = ['N-Gram','Frequency']
    df_trigrams.columns = ['N-Gram', 'Frequency']
    
    df_bigrams[['word1', 'word2']] = df_bigrams['N-Gram'].apply(pd.Series)
    df_trigrams[['word1', 'word2', 'word3']] = df_trigrams['N-Gram'].apply(pd.Series)

    data_dict['Unigrams'] = df_unigrams
    data_dict['Bigrams'] = df_bigrams
    data_dict['Trigrams'] = df_trigrams
    
    return data_dict

def viz(chart_data, num_grams=30, ascending=False, font_scale=1, 
        chart_scale='poster', chart_size=15, save_to_file=True, title=""):
   
    chart_data = chart_data.head(num_grams)\
    .reset_index().sort_values(by='Frequency', ascending=ascending).copy()
    
    sns.set_context(chart_scale, font_scale=font_scale)
    
    sns_plot = sns.catplot(x="Frequency", y='N-Gram', data=chart_data, kind="bar",\
                           size=chart_size, palette="flare");
    plt.title(title)
    plt.show();
    if save_to_file is not False:
        sns_plot.savefig(table_name)
    

titles = ['Unigrams', 'Bigrams', 'Trigrams']

for title in titles:
    viz(data_dict[title].sort_values(by='Frequency', ascending=False), num_grams=10, \
       chart_scale='poster', chart_size=9, title=title)
