==> Project_5/05_Arabic_Topic_Modeling.ipynb <==
import pandas as pd
import numpy as np
import pickle

from sklearn.decomposition import TruncatedSVD
from sklearn.decomposition import NMF
from sklearn.metrics.pairwise import cosine_similarity

ar_dtm = pickle.load(open("tfidf1_sparse.pkl", 'rb'))

ar_vectorizer = pickle.load(open("ar_vectorizer.pkl", 'rb'))

ar_nmf_model = NMF(10)
ar_topics = ar_nmf_model.fit_transform(ar_dtm)
ar_topics.shape

n_top_words = 15
ar_feature_names = ar_vectorizer.get_feature_names()

for i, topic_vec in enumerate(ar_nmf_model.components_):
    print(i, end= ' ')
    for fid in topic_vec.argsort()[-1:-n_top_words-1:-1]:
        print(ar_feature_names[fid], end=' ')
    print()

topic = ar_nmf_model.components_.argsort(axis=1)[:, -1:-9:-1]
topic_words_list = [[ar_feature_names[e] for e in l] for l in topic]
topic_words_list

ar_df['topic_full'] = ar_topics.argmax(axis=1)
ar_df['topic_nouns'] = ar_nouns_topics.argmax(axis=1)
