==> Project_5/06_English_Topic_Modeling.ipynb <==
import pandas as pd
import numpy as np
import pickle

from sklearn.decomposition import TruncatedSVD
from sklearn.decomposition import NMF
from sklearn.metrics.pairwise import cosine_similarity

en_dtm = pickle.load(open("tfidf_en_sparse.pkl", 'rb'))

en_vectorizer = pickle.load(open("en_vectorizer.pkl", 'rb'))

en_nmf_model = NMF(10)
en_topics = en_nmf_model.fit_transform(en_dtm)
en_topics.shape

n_top_words = 15
en_feature_names = en_vectorizer.get_feature_names()

for i, topic_vec in enumerate(en_nmf_model.components_):
    print(i, end= ' ')
    for fid in topic_vec.argsort()[-1:-n_top_words-1:-1]:
        print(en_feature_names[fid], end=' ')
    print()

topic = en_nmf_model.components_.argsort(axis=1)[:, -1:-9:-1]
topic_words_list = [[en_feature_names[e] for e in l] for l in t]
topic_words_list

en_df['topic_full'] = en_topics.argmax(axis=1)


en_df['topic_nouns'] = en_nouns_topics.argmax(axis=1)
