==> Project_5/09_CorEx_English.ipynb <==
import pandas as pd
import numpy as np
import scipy.sparse as ss
import pickle

from corextopic import corextopic as ct
from corextopic import vis_topic as vt

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

from sklearn.feature_extraction.text import CountVectorizer

en_df = pickle.load(open("en_df_3mo.pkl", 'rb'))

en_df.head()

en_df.filtered = en_df.filtered.apply(lambda x: ' '.join(x))

en_df.filtered.head()

en_df.columns

en_df.lang.value_counts()

en_df.head()

en_dtm = pickle.load(open("tfidf_en_3mo_sparse.pkl", 'rb'))

en_vectorizer = pickle.load(open("en_vectorizer_3mo.pkl", 'rb'))

words = list(np.asarray(en_vectorizer.get_feature_names()))

topic_model = ct.Corex(n_hidden=30, words=words, seed=55)
topic_model.fit(en_dtm, words=words, docs=en_df.filtered_joined)

### What Are My Topics?

topics = topic_model.get_topics()
for n, topic in enumerate(topics):
    topic_words,_ = zip(*topic)
    print('{}: '.format(n) + ','.join(topic_words))

### Top Tweets Associated With A Topic

topic_model.get_top_docs(topic=4, n_docs=2)

### Which Topics Has CorEx Assigned Each Document?

anchor_words = [['fire', 'fauci', 'fraud'],
                ['bill', 'gates', 'bioterrorist'], 
                ['honey', 'lemon', 'vitamin','garlic'], 
                ['biological', 'warfare'], 
                ['electromagnetic'], 
                ['deep', 'state', 'new', 'world', 'order'], 
                ['id','2020'], 
                ['vegetarian'], 
                ['herbal'], 
                ['scamdemic', 'plandemic'], 
                ['dirty', 'vaccine'], 
                ['wuflu', 'chinavirus']]

anchors_combo = ['fire', 'fauci', 'fraud',
                'bill', 'gates', 'bioterrorist', 
                'honey', 'lemon', 'vitamin','garlic', 
                'biological', 'warfare', 
                'electromagnetic', 
                'deep', 'state', 'new', 'world', 'order', 
                'id','2020', 'fraud',
                'vegetarian', 'bioterrorist',
                'herbal', 'covidhoax',
                'scamdemic', 'plandemic', 
                'dirty', 'coronavirushoax',
                'wuflu', 'chinavirus']

an_topic_model = ct.Corex(n_hidden=30, max_iter=2000, words=words, seed=55)
an_topic_model.fit(en_dtm, words=words, anchors=anchors_combo, anchor_strength=10, docs=en_df.filtered_joined)

an_topics = an_topic_model.get_topics()
for n, topic in enumerate(an_topics):
    topic_words,_ = zip(*an_topic_model.get_topics(topic=n))
    print('{}: '.format(n) + ','.join(topic_words))

from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords

nltk.download('stopwords')

stop_words_standard = stopwords.words('english')
stop_words_long = open('/Users/jess/workspace/Metis_Projects/Project_4/stopwords/stopwords_long_ranks_nl.txt', 'r', encoding='utf-8').read().split('\n')
stop = set(stop_words_standard + stop_words_long + ['https' , 'http'])

cv = CountVectorizer(lowercase=True,
                     stop_words=stop,
                     min_df = 2,
                     max_df = .8,
                     ngram_range=(1,2)
                    )

corex_vector = cv.fit_transform(en_df.filtered_joined)

cv_topic_model = ct.Corex(n_hidden=30, seed=55, max_iter=3000)
cv_topic_model.fit(corex_vector, anchors=anchor_words, anchor_strength=10,words=cv.get_feature_names())

cv_topics = cv_topic_model.get_topics()
for n, topic in enumerate(cv_topics):
    topic_words,_ = zip(*topic)
    print('{}: '.format(n) + ','.join(topic_words))

w2v_anchors = [['plandemic', 'scamdemic', 'fearmongering', 'coronavirushoax', 'covidhoax', 'fabricated', 'smokescreen']]

import nltk
from nltk.corpus import stopwords
nltk.download('stopwords')

stop_words_standard = stopwords.words('english')
stop_words_long = open('/Users/jess/workspace/Metis_Projects/Project_4/stopwords/stopwords_long_ranks_nl.txt', 'r', encoding='utf-8').read().split('\n')
stop = set(stop_words_standard + stop_words_long + ['https' , 'http'])

texts = [[word.lower() for word in tweet if word not in stop] for tweet in en_df.tokens.values]
print(texts[0:5])
